feat: object storage support through aws s3 sdk

+++ src/main/java/edu/dadaev/greenpoint/dto/ResourceRequestDTO.java
@@ -9,5 +9,5 @@ import java.math.BigDecimal;
 /**
  * DTO for {@link Resource}
  */
-public record ResourceRequestDTO(String name, BigDecimal price, Integer amount, MultipartFile image) implements Serializable {
+public record ResourceRequestDTO(String name, BigDecimal price, BigDecimal amount, MultipartFile image) implements Serializable {
 }
